requirements.txt
numpy==1.22.3
torch==1.11.0
scikit-allel==1.3.5
